LotusOS : CustomSeekBar: fix progressbar glitch with custom min-max values
when CustomSeekBar is called, it does:
mSeekBar.setMax(mMax - mMin);
assuming mMin=0 and mMax=100 if they are not set with settings:min or max
through the related menu xml.
So if we set one of those values in the related menu java class with the
available public voids (setMax(value) and setMin(value)) the seekbar doesn't
get the new min and max thus the user can't slide it to the max position.

To reply:
- set settings:max (or android:max according to your customseekbar implementation)
to the wanted value, e.g. 255;
- set the min value through your menu java class with CustomSeekBarPreference.setMin(value);
- try to move the progressbar thumb to the max position, it won't reach it.

Signed-off-by: xyyx <xyyx@mail.ru>